pyproject.toml
[tool.poetry]
name = "modelserver"
version = "0.2.0"
description = ""
authors = ["Andrew Duffy <andrew@intrinsiclabs.ai>"]
readme = "README.md"

[tool.poetry.scripts]
modelserver = "modelserver.app:entrypoint"
workerserver = "workerserver.worker:entrypoint"
selfsign = "scripts.gen_certs:entrypoint"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.5.0"
fastapi = "^0.104.1"
llama-cpp-python = "^0.2.18"
huggingface-hub = "0.17.3"
duckdb = "^0.9.1"
uvicorn = "^0.23.2"
colorlog = "^6.7.0"
sqlalchemy = "^2.0.23"
websockets = "^11.0.3"
python-dotenv = "^1.0.0"
grpcio = "^1.59.2"
python-multipart = "^0.0.6"


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.0"
mypy = "^1.7.0"
mypy-extensions = "^1.0.0"
isort = "^5.12.0"
black = "^23.11.0"
grpcio-tools = "^1.59.2"
pyopenssl = "^23.3.0"
pytest = "^7.4.3"


[tool.poetry.group.remoteworker.dependencies]
trl = "^0.7.2"
transformers = "^4.35.0"
httpx = "^0.25.1"
peft = "^0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
